# specs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "AI4I-Core",
      "description": "A comprehensive microservices based platform to handle language based AI models inferences at scale.",
      "version": "1.0.0"
    },
    "paths": {
...
      "/api/v1/auth/register": {
        "post": {
          "tags": [
            "Authentication"
          ],
          "summary": "Register User",
          "description": "Register a new user",
          "operationId": "register_user_api_v1_auth_register_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterUser"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/auth/login": {
        "post": {
          "tags": [
            "Authentication"
          ],
          "summary": "Login User",
          "description": "Login user",
          "operationId": "login_user_api_v1_auth_login_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginRequestBody"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/auth/logout": {
        "post": {
          "tags": [
            "Authentication"
          ],
          "summary": "Logout User",
          "description": "Logout user",
          "operationId": "logout_user_api_v1_auth_logout_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutBody"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "$ref": "#/components/parameters/XAuthSource"
            }
          ]
        }
      },
      "/api/v1/auth/refresh": {
        "post": {
          "tags": [
            "Authentication"
          ],
          "summary": "Refresh Token",
          "description": "Refresh access token",
          "operationId": "refresh_token_api_v1_auth_refresh_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenRefreshBody"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/v1/auth/change-password": {
        "post": {
          "tags": [
            "Authentication"
          ],
          "summary": "Change Password",
          "description": "Change the current user's password. Requires:\n- **current_password**: Your current password for verification\n- **new_password**: Your new password (minimum 8 characters, must be strong)\n- **confirm_password**: Confirmation of new password (must match new_password)",
          "operationId": "change_password_api_v1_auth_change_password_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordChangeBody"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "$ref": "#/components/parameters/XAuthSource"
            }
          ]
        }
      },
...
      "/api/v1/auth/reset-password": {
        "post": {
          "tags": [
            "Authentication"
          ],
          "summary": "Reset Password",
          "description": "Reset password with token\n\nCompletes the password reset process using the token received via email.\n\n**How it works**:\n1. User clicks the reset link in their email\n2. Link contains a token (in URL or entered manually)\n3. User provides: token, new password, and confirmation\n4. System validates the token and updates the password\n\n**Parameters**:\n- **token**: The password reset token received via email\n- **new_password**: Your new password (minimum 8 characters, must be strong)\n- **confirm_password**: Confirmation of new password (must match new_password)\n\n**Note**: After successful reset, all existing sessions are invalidated for security.",
          "operationId": "reset_password_api_v1_auth_reset_password_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetConfirmBody"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "LoginRequestBody": {
          "properties": {
            "email": {
              "type": "string",
              "title": "Email",
              "description": "Email address"
            },
            "password": {
              "type": "string",
              "title": "Password",
              "description": "User password"
            },
            "remember_me": {
              "type": "boolean",
              "title": "Remember Me",
              "description": "Issue long-lived refresh token",
              "default": false
            }
          },
          "type": "object",
          "required": [
            "email",
            "password"
          ],
          "title": "LoginRequestBody"
        },
        "LogoutBody": {
          "properties": {
            "refresh_token": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Refresh Token",
              "description": "Refresh token to invalidate; if omitted, logs out all sessions"
            }
          },
          "type": "object",
          "title": "LogoutBody"
        },
...
        "PasswordChangeBody": {
          "properties": {
            "current_password": {
              "type": "string",
              "title": "Current Password",
              "description": "Current password"
            },
            "new_password": {
              "type": "string",
              "maxLength": 100,
              "minLength": 8,
              "title": "New Password",
              "description": "New password (minimum 8 characters)"
            },
            "confirm_password": {
              "type": "string",
              "maxLength": 100,
              "minLength": 8,
              "title": "Confirm Password",
              "description": "Confirm new password (must match new_password)"
            }
          },
          "type": "object",
          "required": [
            "current_password",
            "new_password",
            "confirm_password"
          ],
          "title": "PasswordChangeBody"
        },
        "PasswordResetConfirmBody": {
          "properties": {
            "token": {
              "type": "string",
              "title": "Token",
              "description": "Password reset token received via email"
            },
            "new_password": {
              "type": "string",
              "maxLength": 100,
              "minLength": 8,
              "title": "New Password",
              "description": "New password (minimum 8 characters)"
            },
            "confirm_password": {
              "type": "string",
              "maxLength": 100,
              "minLength": 8,
              "title": "Confirm Password",
              "description": "Confirm new password (must match new_password)"
            }
          },
          "type": "object",
          "required": [
            "token",
            "new_password",
            "confirm_password"
          ],
          "title": "PasswordResetConfirmBody"
        },
...
        "RegisterUser": {
          "properties": {
            "email": {
              "type": "string",
              "title": "Email",
              "description": "Email address"
            },
            "username": {
              "type": "string",
              "maxLength": 100,
              "minLength": 3,
              "title": "Username",
              "description": "Username (3-100 characters)"
            },
            "password": {
              "type": "string",
              "maxLength": 100,
              "minLength": 8,
              "title": "Password",
              "description": "Password (minimum 8 characters)"
            },
            "confirm_password": {
              "type": "string",
              "maxLength": 100,
              "minLength": 8,
              "title": "Confirm Password",
              "description": "Password confirmation (must match password)"
            },
            "full_name": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Full Name",
              "description": "Full name"
            },
            "phone_number": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Phone Number",
              "description": "Phone number"
            },
            "timezone": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Timezone",
              "description": "Timezone",
              "default": "UTC"
            },
            "language": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language",
              "description": "Language code",
              "default": "en"
            }
          },
          "type": "object",
          "required": [
            "email",
            "username",
            "password",
            "confirm_password"
          ],
          "title": "RegisterUser"
        },
...
        "TokenRefreshBody": {
          "properties": {
            "refresh_token": {
              "type": "string",
              "title": "Refresh Token",
              "description": "Refresh token"
            }
          },
          "type": "object",
          "required": [
            "refresh_token"
          ],
          "title": "TokenRefreshBody"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        },
...
      },
      "securitySchemes": {
        "HTTPBearer": {
          "type": "http",
          "scheme": "bearer"
        },
        "APIKeyHeader": {
          "type": "apiKey",
          "in": "header",
          "name": "X-API-Key"
        }
      },
      "parameters": {
        "XAuthSource": {
          "name": "x-auth-source",
          "in": "header",
          "required": false,
          "description": "Select auth source for testing in Swagger (API_KEY or AUTH_TOKEN)",
          "schema": {
            "type": "string",
            "enum": [
              "API_KEY",
              "AUTH_TOKEN"
            ],
            "default": "API_KEY"
          }
        }
      }
    },
    "servers": [
      {
        "url": "http://localhost:8080",
        "description": "API Gateway (with API Key Permission Validation)"
      }
    ],
...
  }